+++ example/lib/main.dart
@@ -19,6 +19,7 @@ import 'screens/inputs/button.dart';
 import 'screens/inputs/checkbox.dart';
 import 'screens/inputs/slider.dart';
 import 'screens/inputs/toggle_switch.dart';
+import 'screens/navigation/navigation_view.dart';
 import 'screens/navigation/tab_view.dart';
 import 'screens/navigation/tree_view.dart';
 import 'screens/settings.dart';
@@ -32,6 +33,7 @@ import 'screens/surface/progress_indicators.dart';
 import 'screens/surface/tooltip.dart';
 import 'screens/theming/colors.dart';
 import 'screens/theming/icons.dart';
+import 'screens/theming/reveal_focus.dart';
 import 'screens/theming/typography.dart';
 import 'theme.dart';
 
@@ -249,6 +251,10 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
       icon: const Icon(FluentIcons.icon_sets_flag),
       title: const Text('Icons'),
     ),
+    PaneItem(
+      icon: const Icon(FluentIcons.focus),
+      title: const Text('Reveal Focus'),
+    ),
   ];
   final List<NavigationPaneItem> footerItems = [
     PaneItemSeparator(),
@@ -278,8 +284,7 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
     TimePickerPage(),
     DatePickerPage(),
     // navigation
-    EmptyPage(), // navigation view
-
+    NavigationViewPage(),
     TabViewPage(),
     TreeViewPage(),
     // surfaces
@@ -296,8 +301,11 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
     ColorsPage(),
     const TypographyPage().toPage(),
     const IconsPage().toPage(),
+    RevealFocusPage(),
     // others
     Settings(),
+
+    // TODO: mobile widgets, Scrollbar, BottomNavigationBar, RatingBar
   ];
 
   @override