+++ example/lib/screens/home.dart
@@ -3,6 +3,7 @@ import 'package:url_launcher/link.dart';
 
 import '../models/sponsor.dart';
 import '../widgets/changelog.dart';
+import '../widgets/material_equivalents.dart';
 import '../widgets/page.dart';
 import '../widgets/sponsor.dart';
 
@@ -60,12 +61,13 @@ class HomePage extends ScrollablePage {
               ),
             ),
           ),
-          const RepaintBoundary(
+          RepaintBoundary(
             child: Padding(
-              padding: EdgeInsetsDirectional.only(start: 4.0),
+              padding: const EdgeInsetsDirectional.only(start: 4.0),
               child: InfoLabel(
                 label: 'Progress',
-                child: SizedBox(height: 30, width: 30, child: ProgressRing()),
+                child: const SizedBox(
+                    height: 30, width: 30, child: ProgressRing()),
               ),
             ),
           ),
@@ -88,9 +90,9 @@ class HomePage extends ScrollablePage {
               ]),
             ),
           ),
-          const InfoLabel(
+          InfoLabel(
             label: 'Icons',
-            child: Icon(FluentIcons.graph_symbol, size: 30.0),
+            child: const Icon(FluentIcons.graph_symbol, size: 30.0),
           ),
           InfoLabel(
             label: 'Colors',
@@ -208,9 +210,11 @@ class HomePage extends ScrollablePage {
               ),
               const Text('Become a Sponsor!'),
             ]),
-          )
+          ),
         ],
       ),
+      subtitle(content: const Text('Equivalents with the material library')),
+      const MaterialEquivalents(),
     ];
   }
 }